refactor with new aircraft package
