+++ app/src/main/java/org/WenuLink/controllers/MAVLinkController.kt
@@ -34,10 +34,10 @@ import com.MAVLink.enums.MAV_RESULT
 import com.MAVLink.minimal.msg_heartbeat
 import io.getstream.log.taggedLogger
 import kotlin.math.roundToInt
-import org.WenuLink.adapters.AircraftHandler
 import org.WenuLink.adapters.AsyncUtils
-import org.WenuLink.adapters.MessageRate
 import org.WenuLink.adapters.MessageUtils
+import org.WenuLink.adapters.aircraft.AircraftHandler
+import org.WenuLink.adapters.aircraft.MessageRate
 import org.WenuLink.mavlink.MAVLinkClient
 
 /**
